Split-out Slack user and channel management
